modifications of modules to avoid circular imports
