+++ src/pymodaq/control_modules/daq_viewer.py
@@ -45,8 +45,8 @@ from pymodaq_gui.utils.utils import mkQApp
 
 from pymodaq.utils.gui_utils import get_splash_sc
 from pymodaq.control_modules.daq_viewer_ui import DAQ_Viewer_UI
-from pymodaq.control_modules.utils import (DET_TYPES, get_viewer_plugins, DAQTypesEnum,
-                                           DetectorError)
+from pymodaq.control_modules.instruments import (DET_TYPES, DAQTypesEnum,
+                                           DetectorError, get_viewer_plugins)
 from pymodaq.control_modules.thread_commands import (ThreadStatus, ThreadStatusViewer, ControlToHardwareViewer,
                                                      UiToMainViewer)
 from pymodaq_gui.plotting.data_viewers.viewer import ViewerBase
