+++ src/pymodaq/control_modules/daq_viewer_ui.py
@@ -18,7 +18,7 @@ from pymodaq.control_modules.ui_utils import ControlModuleUI
 from pymodaq_gui.utils.widgets import PushButtonIcon, LabelWithFont, QLED
 from pymodaq_gui.utils import Dock
 from pymodaq_utils.config import Config as ConfigUtils
-from pymodaq.control_modules.utils import DET_TYPES, DAQTypesEnum
+from pymodaq.control_modules.instruments import DET_TYPES, DAQTypesEnum
 from pymodaq_gui.plotting.data_viewers.viewer import ViewerFactory, ViewerDispatcher
 from pymodaq_gui.plotting.data_viewers import ViewersEnum
 from pymodaq_utils.enums import enum_checker
